implement query when fetching services
